feat: add ThemeProvider to top lvl
