+++ src/App.tsx
@@ -1,22 +1,36 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { ThemeProvider, createTheme } from '@mui/material/styles';
 import PageStart from './pages/Start';
 import PageAbout from './pages/About';
 import PageHome from './pages/Home';
 import PageNotFound from './pages/NotFound';
 import PageHistory from './pages/History';
 
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#000',
+    },
+    secondary: {
+      main: '#000',
+    },
+  },
+});
+
 const App = () => {
   return (
     <>
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<PageStart />} />
-          <Route path="/home" element={<PageHome />} />
-          <Route path="/about" element={<PageAbout />} />
-          <Route path="/history" element={<PageHistory />} />
-          <Route path="*" element={<PageNotFound />} />
-        </Routes>
-      </BrowserRouter>
+      <ThemeProvider theme={theme}>
+        <BrowserRouter>
+          <Routes>
+            <Route path="/" element={<PageStart />} />
+            <Route path="/home" element={<PageHome />} />
+            <Route path="/about" element={<PageAbout />} />
+            <Route path="/history" element={<PageHistory />} />
+            <Route path="*" element={<PageNotFound />} />
+          </Routes>
+        </BrowserRouter>
+      </ThemeProvider>
     </>
   );
 };
